Use type float32 instead of float64 for regularization weights.

+++ cpp/subprojects/boosting/include/mlrl/boosting/rule_evaluation/rule_evaluation_decomposable_single.hpp
@@ -16,9 +16,9 @@ namespace boosting {
     class DecomposableSingleOutputRuleEvaluationFactory final : public ISparseDecomposableRuleEvaluationFactory {
         private:
 
-            const float64 l1RegularizationWeight_;
+            const float32 l1RegularizationWeight_;
 
-            const float64 l2RegularizationWeight_;
+            const float32 l2RegularizationWeight_;
 
         public:
 
@@ -28,8 +28,8 @@ namespace boosting {
              * @param l2RegularizationWeight    The weight of the L2 regularization that is applied for calculating the
              *                                  scores to be predicted by rules
              */
-            DecomposableSingleOutputRuleEvaluationFactory(float64 l1RegularizationWeight,
-                                                          float64 l2RegularizationWeight);
+            DecomposableSingleOutputRuleEvaluationFactory(float32 l1RegularizationWeight,
+                                                          float32 l2RegularizationWeight);
 
             std::unique_ptr<IRuleEvaluation<DenseDecomposableStatisticVector<float64>>> create(
               const DenseDecomposableStatisticVector<float64>& statisticVector,
